+++ .storybook/.babelrc
@@ -1,7 +1,10 @@
 {
   "presets": [
-    "env",
-    "react"
+    "@babel/preset-env",
+    "@babel/preset-react"
   ],
-  "plugins": ["transform-decorators-legacy"]
+  "plugins": [
+    ["@babel/plugin-proposal-decorators", { "legacy": true} ],
+    ["@babel/plugin-proposal-class-properties", { "loose": true} ]
+  ]
 }
