+++ docs/index.html
@@ -8,5 +8,5 @@
 </head>
 <body>
     <div id="oc-examples"></div>
-<script type="text/javascript" src="vendor.bundle.js"></script><script type="text/javascript" src="examples.js"></script></body>
+<script type="text/javascript" src="examples.js"></script></body>
 </html>
