+++ package.json
@@ -26,7 +26,8 @@
     "preversion": "npm run lint && npm test",
     "version": "npm run build && git add -A lib docs",
     "postversion": "git push && git push --tags",
-    "storybook": "start-storybook -p 9001 -c .storybook"
+    "storybook:hot": "start-storybook -p 6060",
+    "storybook:build": "build-storybook"
   },
   "files": [
     "/lib/"
@@ -42,79 +43,84 @@
     "styled-components": "4"
   },
   "devDependencies": {
-    "@storybook/addon-actions": "3.4.8",
-    "@storybook/addon-knobs": "3.4.8",
-    "@storybook/addon-links": "3.4.7",
-    "@storybook/addon-options": "3.4.8",
-    "@storybook/react": "3.4.7",
-    "autoprefixer": "9.4.7",
-    "babel-cli": "6.26.0",
+    "@babel/cli": "7.2.3",
+    "@babel/core": "7.3.4",
+    "@babel/plugin-proposal-class-properties": "7.3.4",
+    "@babel/plugin-proposal-decorators": "7.3.0",
+    "@babel/plugin-proposal-export-default-from": "7.2.0",
+    "@babel/plugin-proposal-export-namespace-from": "7.2.0",
+    "@babel/plugin-syntax-dynamic-import": "7.2.0",
+    "@babel/polyfill": "7.2.5",
+    "@babel/preset-env": "7.3.4",
+    "@babel/preset-react": "7.0.0",
+    "@babel/register": "7.0.0",
+    "@storybook/addon-actions": "5.0.11",
+    "@storybook/addon-knobs": "5.0.11",
+    "@storybook/addon-options": "5.0.11",
+    "@storybook/react": "5.0.11",
+    "autoprefixer": "9.5.1",
     "babel-eslint": "10.0.1",
-    "babel-loader": "7.1.5",
+    "babel-loader": "8.0.5",
     "babel-plugin-dynamic-import-node": "2.2.0",
-    "babel-plugin-react-transform": "3.0.0",
-    "babel-plugin-transform-decorators-legacy": "1.3.5",
-    "babel-plugin-transform-react-remove-prop-types": "0.4.23",
-    "babel-polyfill": "6.26.0",
-    "babel-preset-env": "1.7.0",
-    "babel-preset-react": "6.24.1",
-    "babel-preset-stage-1": "6.24.1",
-    "babel-register": "6.26.0",
-    "bootstrap-sass": "3.4.0",
+    "babel-plugin-transform-react-remove-prop-types": "0.4.24",
+    "bootstrap-sass": "3.4.1",
     "chai": "4.2.0",
     "chai-immutable": "1.6.0",
-    "clean-webpack-plugin": "1.0.1",
+    "clean-webpack-plugin": "2.0.2",
     "cross-env": "5.2.0",
-    "css-loader": "0.28.11",
-    "enzyme": "3.8.0",
-    "enzyme-adapter-react-16": "1.8.0",
-    "eslint": "5.12.1",
+    "css-loader": "2.1.1",
+    "enzyme": "3.9.0",
+    "enzyme-adapter-react-16": "1.12.1",
+    "eslint": "5.16.0",
     "eslint-config-airbnb": "16.1.0",
     "eslint-plugin-import": "2.16.0",
-    "eslint-plugin-jsx-a11y": "6.2.0",
+    "eslint-plugin-jsx-a11y": "6.2.1",
     "eslint-plugin-react": "7.12.4",
-    "file-loader": "2.0.0",
+    "file-loader": "3.0.1",
     "global-jsdom": "4.2.0",
-    "html-webpack-plugin": "2.30.1",
+    "html-webpack-plugin": "3.2.0",
     "ignore-styles": "5.0.1",
     "immutable": "3.8.2",
-    "jsdom": "13.2.0",
-    "mocha": "5.2.0",
+    "jsdom": "15.0.0",
+    "mini-css-extract-plugin": "0.5.0",
+    "mocha": "6.1.4",
     "moment": "2.23.0",
     "node-sass": "4.11.0",
     "npm-run-all": "4.1.5",
+    "optimize-css-assets-webpack-plugin": "5.0.1",
     "postcss-flexbugs-fixes": "4.1.0",
-    "postcss-loader": "2.1.6",
+    "postcss-loader": "3.0.0",
     "precss": "4.0.0",
     "progress-bar-webpack-plugin": "1.12.1",
-    "prop-types": "15.6.2",
+    "prop-types": "15.7.2",
     "raf": "3.4.1",
     "react": "16.6.3",
     "react-bootstrap": "0.32.4",
     "react-dom": "16.6.3",
-    "react-hot-loader": "4.3.11",
+    "react-hot-loader": "4.8.4",
     "react-router": "4.3.1",
     "react-router-dom": "4.3.1",
     "react-svg-loader": "2.1.0",
-    "rimraf": "2.6.2",
+    "rimraf": "2.6.3",
     "sass-loader": "7.1.0",
-    "sinon": "7.2.3",
+    "sinon": "7.3.2",
     "style-loader": "0.23.1",
     "styled-components": "4.1.3",
     "url-loader": "1.1.2",
     "uuid": "3.3.2",
-    "webpack": "3.12.0",
-    "webpack-dev-server": "2.11.3",
-    "webpack-merge": "4.1.5",
+    "webpack": "4.30.0",
+    "webpack-cli": "3.3.2",
+    "webpack-dev-server": "3.3.1",
+    "webpack-merge": "4.2.1",
     "webpack-node-externals": "1.7.2",
     "webpack-notifier": "1.7.0",
     "write-file-webpack-plugin": "4.5.0"
   },
   "dependencies": {
-    "@opuscapita/oc-cm-common-layouts": "3.0.0",
+    "@opuscapita/oc-cm-common-layouts": "3.1.0",
     "@opuscapita/oc-cm-common-styles": "3.0.0",
-    "@opuscapita/react-datetime": "3.2.12",
-    "@opuscapita/react-floating-select": "3.3.2"
+    "@opuscapita/react-datetime": "3.3.2",
+    "@opuscapita/react-floating-select": "3.6.1"
   },
   "repository": {
     "type": "git",
