+++ webpack.docs.config.js
@@ -4,9 +4,12 @@ const merge = require('webpack-merge');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const WebpackNotifierPlugin = require('webpack-notifier');
 const CleanWebpackPlugin = require('clean-webpack-plugin');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 const autoprefixer = require('autoprefixer');
 const precss = require('precss');
 const flexbugs = require('postcss-flexbugs-fixes');
+const packageConfig = require('./package.json');
 
 const isProd = process.env.NODE_ENV === 'production';
 
@@ -14,6 +17,9 @@ const PATHS = {
   build: path.join(__dirname, 'docs'),
   context: path.join(__dirname, 'src_docs'),
   jsFileName: 'examples.js',
+  jsChunkFileName: 'examples.[name].js',
+  cssFileName: 'examples.css',
+  cssChunkFileName: 'examples.[name].css',
   entry: path.join(__dirname, 'src_docs', 'index.jsx'),
   root: __dirname,
 };
@@ -24,12 +30,13 @@ const PATHS = {
 const baseConfig = {
   context: PATHS.context,
   entry: [
-    'babel-polyfill',
+    '@babel/polyfill',
     PATHS.entry,
   ],
   output: {
     path: PATHS.build,
     filename: PATHS.jsFileName,
+    chunkFilename: PATHS.jsChunkFileName,
   },
   module: {
     rules: [
@@ -47,56 +54,21 @@ const baseConfig = {
         ],
       },
       {
-        test: /\.css$/,
-        use: [
-          'style-loader',
-          'css-loader',
-          {
-            loader: 'postcss-loader',
-            options: {
-              plugins: () => [flexbugs, precss, autoprefixer],
-            },
-          },
-        ],
-      },
-      {
-        test: /\.scss$/,
-        use: [
-          'style-loader',
-          'css-loader',
-          {
-            loader: 'postcss-loader',
-            options: {
-              plugins: () => [flexbugs, precss, autoprefixer],
-            },
-          },
-          'sass-loader',
-        ],
-      },
-      {
         test: /\.svg$/,
-        exclude: path.resolve(__dirname, 'node_modules', 'font-awesome'),
-        use: ['babel-loader', 'react-svg-loader'],
-      },
-      {
-        test: /\.svg$/,
-        include: path.resolve(__dirname, 'node_modules', 'font-awesome'),
         use: [{
-          loader: 'file-loader',
-          options: {
-            name: '[name].[ext]',
-            outputPath: 'fonts/',
-          },
+          loader: 'babel-loader',
+        },
+        {
+          loader: 'react-svg-loader',
         }],
       },
       {
         test: /\.(woff|woff2)(\?v=\d+\.\d+\.\d+)?$/,
         use: [{
-          loader: 'url-loader',
+          loader: 'file-loader',
           options: {
             name: '[name].[ext]',
             outputPath: 'fonts/',
-            limit: 100,
             mimetype: 'application/font-woff',
           },
         }],
@@ -104,11 +76,10 @@ const baseConfig = {
       {
         test: /\.ttf(\?v=\d+\.\d+\.\d+)?$/,
         use: [{
-          loader: 'url-loader',
+          loader: 'file-loader',
           options: {
             name: '[name].[ext]',
             outputPath: 'fonts/',
-            limit: 100,
             mimetype: 'application/octet-stream',
           },
         }],
@@ -138,10 +109,8 @@ const baseConfig = {
     fs: 'empty',
   },
   plugins: [
-    new CleanWebpackPlugin([PATHS.build], {
-      root: PATHS.root,
+    new CleanWebpackPlugin({
       verbose: true,
-      dry: false,
     }),
     new HtmlWebpackPlugin({
       filename: 'index.html',
@@ -162,15 +131,36 @@ const baseConfig = {
 * DEVELOPMENT CONFIG
 */
 const devConfig = {
+  mode: 'development',
   devtool: 'eval-source-map',
+  module: {
+    rules: [
+      {
+        test: /\.(sa|sc|c)ss$/,
+        use: [
+          'style-loader',
+          'css-loader',
+          {
+            loader: 'postcss-loader',
+            options: {
+              plugins: () => [flexbugs, precss, autoprefixer],
+            },
+          },
+          'sass-loader',
+        ],
+      },
+    ],
+  },
   plugins: [
     new webpack.DefinePlugin({
       'process.env': {
         NODE_ENV: JSON.stringify('development'),
       },
     }),
-    new WebpackNotifierPlugin(),
-    new webpack.NamedModulesPlugin(),
+    new WebpackNotifierPlugin({
+      title: packageConfig.name,
+      contentImage: path.join(__dirname, 'src_docs', 'images', 'favicon.ico'),
+    }),
   ],
 };
 
@@ -178,25 +168,43 @@ const devConfig = {
 * PRODUCTION CONFIG
 */
 const prodConfig = {
-  devtool: 'source-map',
+  mode: 'production',
+  devtool: false,
+  module: {
+    rules: [
+      {
+        test: /\.(sa|sc|c)ss$/,
+        use: [
+          MiniCssExtractPlugin.loader,
+          'css-loader',
+          {
+            loader: 'postcss-loader',
+            options: {
+              plugins: () => [flexbugs, precss, autoprefixer],
+            },
+          },
+          'sass-loader',
+        ],
+      },
+    ],
+  },
   plugins: [
-    new webpack.optimize.CommonsChunkPlugin({
-      name: 'vendor',
-      filename: 'vendor.bundle.js',
-    }),
     new webpack.DefinePlugin({
       'process.env': {
         NODE_ENV: JSON.stringify('production'),
       },
     }),
-    new webpack.optimize.ModuleConcatenationPlugin(),
-    new webpack.optimize.UglifyJsPlugin({
-      sourceMap: true,
-      output: {
-        comments: false,
-      },
+    new OptimizeCssAssetsPlugin(),
+    new MiniCssExtractPlugin({
+      filename: PATHS.cssFileName,
+      chunkFilename: PATHS.cssChunkFileName,
     }),
   ],
+  optimization: {
+    splitChunks: {
+      chunks: 'all',
+    },
+  },
 };
 
 module.exports = merge(baseConfig, isProd ? prodConfig : devConfig);
