+++ webpack.docs.hot.config.js
@@ -29,11 +29,4 @@ const hotConfig = {
   },
 };
 
-const mergedConfig = merge(devConfig, hotConfig);
-mergedConfig.entry = [
-  'babel-polyfill',
-  'react-hot-loader/patch',
-  devConfig.entry[1],
-];
-
-module.exports = mergedConfig;
+module.exports = merge(devConfig, hotConfig);
